refactor: decoupled History components into smaller components
